SWPTP-1278: Fix use of msghdr->cmsg_len
This patch wraps the ancillary data size with the CMSG_LEN() macro so
that the resulting size accommodates the control message header and can
be directly compared with msghdr->cmsg_len, which includes the header
size.

Without CMSG_LEN(), we allowed truncated ancillary data up to the
control message header size.