SWPTP-976: try to send DelayRequest out of last Sync port using onload
Some PTP configurations may have a bond connected to multiple boundary
clocks. In these configurations, the bond will hash some of the packet
information to select an output interface. This can result in messages
going to the wrong boundary clock.

This change attempts to mitigate the effects of this failure mode by
requesting to send the packet on the interface that last received a
Sync message.