clock: avoid checking phc state for system clock
clock->u.nic.phc happens typically to be non-zero for a system clock type in
most systems but apparently not on Raspberry Pi 5 with Debian 12. It could be
a layout change or system clock characteristics change but in any case the
code should not have been checking this field. Fixes Xilinx-CNS/sfptpd#9