SWPTP-1476: tolerate no interface in doInitPort instead of crashing
Previously I had falsely assumed doInitPort would always be called
when we had a valid interface to send over, however this was not
always true. It was possible that our only interface could go down
and we would enter PTP_FAULTY, then PTP_INITIALIZING after some
time. If the interface was still not up by this time, we could go
through the following steps:
- ptpd_timer_tick calls doInitInterface
- doInitInterface calls netInit
- netInit calls netInitTimestamping
- The above calls all fail if we have no interface, and
  ptpd_timer_tick continues by calling doInitPort.

This change replaces the assertions on having a valid interface
and clock with conditionals that instead move the PTP state
back into PTP_FAULTY.